git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk@522821 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/broker/src/main/java/org/apache/qpid/server/queue/AMQQueue.java
@@ -118,7 +118,7 @@ public class AMQQueue implements Managable, Comparable
 
     /** max allowed number of messages on a queue. */
     @Configured(path = "maximumMessageCount", defaultValue = "0")
-    public int _maximumMessageCount;
+    public long _maximumMessageCount;
 
     /** max queue depth for the queue */
     @Configured(path = "maximumQueueDepth", defaultValue = "0")
@@ -350,12 +350,12 @@ public class AMQQueue implements Managable, Comparable
         return _totalMessagesReceived.get();
     }
 
-    public int getMaximumMessageCount()
+    public long getMaximumMessageCount()
     {
         return _maximumMessageCount;
     }
 
-    public void setMaximumMessageCount(int value)
+    public void setMaximumMessageCount(long value)
     {
         _maximumMessageCount = value;
     }
